refactored in accordance with PULL review
