+++ types/index.ts
@@ -20,3 +20,19 @@ export type ProfileType = {
 export type ResumeType = {
   resumeURL: string;
 };
+
+export type MobileNavbarProps = {
+  toggleMobile: () => void;
+  showMobileNavbar: boolean;
+  closeImgSrc: string;
+  pathname: string;
+};
+
+export type NavbarRoutesProps = {
+  pathname: string;
+  showMobileNavbar: boolean;
+};
+
+export type DeskTopNavbarProps = Omit<MobileNavbarProps, 'closeImgSrc'> & {
+  openImgSrc: string;
+};
